WIP - refactor again. Add folder for develpment and move batch inference into deployment. Move data inside project folder
