+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/deployment/batch_inference/BatchInference.py
