+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/development/feature_engineering/features/pickup_features.py
