+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/development/feature_engineering/notebooks/GenerateAndWriteFeatures.py
