+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/development/training/notebooks/Train.py
