+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/development/training/profiles/databricks-prod.yaml
