+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/development/training/profiles/local.yaml
