+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/development/training/steps/train.py
