+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/monitoring/README.md
@@ -1,0 +1,3 @@
+# Model Monitoring
+
+MLOps stacks model monitoring support coming soon...
