+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/tests/development/feature_engineering/dropoff_features_test.py
@@ -4,7 +4,7 @@ import pandas as pd
 from datetime import datetime
 from pyspark.sql import SparkSession
 
-from feature_engineering.features.pickup_features import compute_features_fn
+from development.feature_engineering.features.dropoff_features import compute_features_fn
 
 
 @pytest.fixture(scope="session")
@@ -24,12 +24,12 @@ def spark(request):
 
 
 @pytest.mark.usefixtures("spark")
-def test_pickup_features_fn(spark):
+def test_dropoff_features_fn(spark):
     input_df = pd.DataFrame(
         {
-            "tpep_pickup_datetime": [datetime(2022, 1, 12)],
-            "tpep_dropoff_datetime": [datetime(2022, 1, 12)],
-            "pickup_zip": [94400],
+            "tpep_pickup_datetime": [datetime(2022, 1, 10)],
+            "tpep_dropoff_datetime": [datetime(2022, 1, 10)],
+            "dropoff_zip": [94400],
             "trip_distance": [2],
             "fare_amount": [100],
         }
@@ -39,4 +39,4 @@ def test_pickup_features_fn(spark):
         spark_df, "tpep_pickup_datetime", datetime(2022, 1, 1), datetime(2022, 1, 15)
     )
     assert isinstance(output_df, pyspark.sql.DataFrame)
-    assert output_df.count() == 4  # 4 15-min intervals over 1 hr window.
+    assert output_df.count() == 1
