+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/tests/development/training/ingest_test.py
@@ -3,7 +3,7 @@ import os
 import tempfile
 import pandas as pd
 from pandas import DataFrame
-from training.steps.ingest import load_file_as_dataframe
+from development.training.steps.ingest import load_file_as_dataframe
 
 
 @pytest.fixture
