+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/tests/development/training/split_test.py
@@ -2,7 +2,7 @@ import pytest
 import os
 import pandas as pd
 from pandas import DataFrame
-from training.steps.split import process_splits
+from development.training.steps.split import process_splits
 
 
 @pytest.fixture
