+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/tests/development/training/test_sample.parquet
