+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/tests/development/training/train_test.py
@@ -1,4 +1,4 @@
-from training.steps.train import estimator_fn
+from development.training.steps.train import estimator_fn
 from sklearn.utils.estimator_checks import check_estimator
