+++ {{cookiecutter.root_dir}}/{{cookiecutter.project_name}}/tests/development/training/transform_test.py
@@ -1,4 +1,4 @@
-from training.steps.transform import transformer_fn
+from development.training.steps.transform import transformer_fn
 
 
 def test_tranform_fn_returns_object_with_correct_spec():
